Updated API in File namespace [devel]
Summary:
Pushed the last set of 'local' implementations to the OD::FileSystemAccess Factory.
Added a proper API for setting and retrieving the file permissions, based on the available
Qt enum. Most importantely, set a file read-only, writable or executable should no longer be
done using a machine command for each file. Keeping this functionality only for large
recursive operations on directories.

Test Plan:
Most using the test programs, more to be done.
Not tested at all on Windows yet.

Reviewers: arnaud, nanne

Reviewed By: nanne

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com/D17314

+++ src/uiIo/uisurvinfoed.cc
@@ -146,25 +146,15 @@ uiSurveyInfoEditor::uiSurveyInfoEditor( uiParent* p, SurveyInfo& si,
     }
     else
     {
-	BufferString storagedir = FilePath(orgstorepath_).add(orgdirname_)
-							 .fullPath();
-	int linkcount = 0;
-	while ( linkcount++ < 20 && File::isLink(storagedir) )
+	BufferString storagedir =
+		    FilePath(orgstorepath_).add(orgdirname_).fullPath();
+	if ( File::isSymLink(storagedir) )
 	{
-	    BufferString newstoragedir = File::linkTarget(storagedir);
+	    BufferString newstoragedir = File::linkEnd( storagedir );
 	    FilePath fp( newstoragedir );
-	    if ( !fp.isAbsolute() )
-	    {
-		fp.setPath( FilePath(storagedir).pathOnly() );
-		newstoragedir = fp.fullPath();
-	    }
-	    storagedir = newstoragedir;
-	}
-	if ( linkcount < 20 )
-	{
-	    FilePath fp( storagedir );
 	    orgstorepath_ = fp.pathOnly();
 	    orgdirname_ = fp.fileName();
+	    storagedir = newstoragedir;
 	}
 
 	fulldirpath = storagedir;
@@ -698,7 +688,7 @@ bool uiSurveyInfoEditor::rejectOK( CallBacker* )
 	const BufferString dirnm = FilePath(orgstorepath_).add(orgdirname_)
 							  .fullPath();
 	if ( File::exists(dirnm) )
-	    File::remove( dirnm );
+	    File::removeDir( dirnm );
     }
 
     return true;
@@ -824,11 +814,8 @@ bool uiSurveyInfoEditor::handleCurrentSurvey()
 	return false;
 
     BufferString linkpos = FilePath(rootdir_).add(newdirnm).fullPath();
-    if ( File::exists(linkpos) )
-    {
-	if ( File::isLink(linkpos) )
-	    File::remove( linkpos );
-    }
+    if ( File::exists(linkpos) && File::isSymLink(linkpos) )
+	File::remove( linkpos );
 
     if ( !File::exists(linkpos) )
     {
@@ -931,18 +918,15 @@ bool uiSurveyInfoEditor::acceptOK( CallBacker* )
 			    newstorepath,newdirnm) )
 	    return false;
 	else if ( !uiMSG().askGoOn(tr("Keep the survey at the old location?")) )
-	    File::remove( olddir );
+	    File::removeDir( olddir );
     }
     else if ( dirnamechanged && !renameSurv(orgstorepath_,orgdirname_,
 								newdirnm) )
 	    return false;
 
     BufferString linkpos = FilePath(rootdir_).add(newdirnm).fullPath();
-    if ( File::exists(linkpos) )
-    {
-       if ( File::isLink(linkpos) )
-	   File::remove( linkpos );
-    }
+    if ( File::exists(linkpos) && File::isSymLink(linkpos) )
+       File::remove( linkpos );
 
     if ( !File::exists(linkpos) )
     {
